### 2019-05-30 #1482 @whedon
[PRE REVIEW]: citesdb: A high-performance database of shipment-level CITES trade data
**Submitting author:** @noamross (<a href="http://orcid.org/0000-0002-2136-0000">Noam Ross</a>)
**Repository:** <a href="https://github.com/ropensci/citesdb" target ="_blank">https://github.com/ropensci/citesdb</a>
**Version:** 0.2.0
