### 2019-05-30 #1482 @noamross
I note that this submission has passed peer-review at rOpenSci: https://github.com/ropensci/software-review/issues/292
