### 2019-05-30 #1482 @danielskatz
Thanks - this means the process for @kyleniemeyer is much simpler: https://joss.readthedocs.io/en/latest/editing.html#processing-of-ropensci-reviewed-and-accepted-submissions
