### 2019-05-30 #1483 @kyleniemeyer
@noamross it looks like there is an issue with the YAML header in the paper preventing compilation
