### 2019-05-31 #1483 @kyleniemeyer
@noamross ah, one more thing to fix. The affiliations need to be structured like this instead:
```
authors:
 - name: Noam Ross
   orcid: 0000-0002-2136-0000
   affiliation: 1
 - name: Evan A. Eskew
   orcid: 0000-0002-1153-5356
   affiliation: 1
 - name: Nicolas Ray
   affiliation: "2, 3"
```
(the quotes are needed for multiple affiliations)
