### 2019-05-31 #1483 @kyleniemeyer
@noamross can you archive the repo on Zenodo and provide the DOI? (or just the latter if the former has already been done)
