### 2019-05-31 #1483 @kyleniemeyer
@noamross @eveskew it looks like the DOI for the Raasveldt and Muhleisen 2018 reference is not correct. Could you fix this and let me know when you have?
