### 2019-05-31 #1483 @noamross
Review of the source code can be found at https://github.com/ropensci/software-review/issues/292
