### 2019-06-02 #1483 @kyleniemeyer
@noamross congrats on your submission's publication in JOSS, and thanks for sharing that link in here for future reference.
