### 2019-06-03 #1483 @arfon
OK, the PDF is fixed here: https://github.com/openjournals/joss-papers/blob/master/joss.01483/10.21105.joss.01483.pdf

It can take a few hours before the caching of the PDFs expires so you might not see this change reflected on the JOSS site immediately.
